requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
